[day_XX] Add project_options to CMake
Based on the template form Lefticus. Not using dynamic_projects_options
since there is an issue with asan in VS2022 that needs to be
investigated.
Some small fixes have been implemented because the warning level has
been increased and they now produce errors.

+++ src/day_08.cpp
@@ -17,9 +17,9 @@ void scan(const Grid& grid, int rs, int re, int ri, int cs, int ce, int ci, auto
     }
 }
 
-int is_visible(const Grid& grid, size_t r, size_t c)
+int is_visible(const Grid& grid, int r, int c)
 {
-    const auto grid_size = grid.size();
+    const auto grid_size = static_cast<int>(grid.size());
 
     if (r == 0 || r == grid_size - 1 || c == 0 || c == grid_size - 1)
         return 1;
@@ -56,9 +56,9 @@ int is_visible(const Grid& grid, size_t r, size_t c)
     return visible;
 }
 
-int get_scenic_score(const Grid& grid, size_t r, size_t c)
+int get_scenic_score(const Grid& grid, int r, int c)
 {
-    const auto grid_size = grid.size();
+    const auto grid_size = static_cast<int>(grid.size());
     if (r == 0 || r == grid_size - 1 || c == 0 || c == grid_size - 1)
         return 0;
